modulized functions getAtomSelectors and getComponentNames into own files and added testing folder
